Refactor code to provide a tuf package
