+++ tuf/http.go
@@ -1,4 +1,4 @@
-package internal
+package tuf
 
 import (
 	"bytes"
